+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changed ngrok to always restart unless stopped
+- Taught Shopify cog to fuzzy match customers as a plan B
 
 ## [1.0.0] - 2024-09-01